% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCC.R
\name{getCC}
\alias{getCC}
\title{Generate graph and calculate its connected components}
\usage{
getCC(adjM)
}
\arguments{
\item{adjM}{a \code{numerical} \code{matrix} containing the adjacency matrix, with
value >0 or 0 indicating whether or not two proteins are identified by shared
peptide(s)}
}
\value{
a \code{list} of two elements: i. a \code{graph} representing protein-to-protein
connections encoded by the adjacency matrix; ii. a \code{list} of \code{vectors}
(one for each connected component) enumerating protein members of each connected
component.
}
\description{
Build a graph of protein-to-protein connections from adjacency matrix and calculate
its connected components.
}
\examples{
# Read the tab-delimited file containing he proteome incidence matrix
incM_filename <- system.file( "extdata"
                             , "incM_Example"
                             , package = "CCs4prot"
                             , mustWork = TRUE)
rownames_filename <- system.file( "extdata"
                                  , "peptideIDs_incM_Example"
                                  , package = "CCs4prot"
                                  , mustWork = TRUE)
colnames_filename <- system.file( "extdata"
                                 , "proteinIDs_incM_Example"
                                 , package = "CCs4prot"
                                 , mustWork = TRUE)
incM <- readIncM(incM_filename=incM_filename
                 , colnames_filename=colnames_filename
                 , rownames_filename=rownames_filename)
# Only retain proteins with at least one shared peptide and all peptides
# mapping on such proteins.
incM_reduced <- reduceIncM(incM)
# Generate adjacency matrix describing protein-to-protein mappings
adjM <- getAdjM(incM_reduced)
# Generate graph of protein-to-protein connections and calculate its connected
# components
multProteinCC <- getCC(adjM)

}
\author{
Laura Fancello
}
